Move visibility into its own concern
